Refactor: Adjusted AetherMapColorUtil to not crash on servers
