+++ gradle.properties
@@ -3,7 +3,7 @@
 org.gradle.jvmargs=-Xmx2G
 max_errors=400
 # Mod Properties
-mod_version=1.6.5+1.18.1
+mod_version=1.6.5.1+1.18.1
 maven_group=com.aether
 archives_base_name=paradise-lost
 # Fabric Properties
